requirements.txt
pillow==10.4.0
torch==2.4.1
torchaudio==2.4.1
torchvision==0.20.0
tqdm==4.67.1
scikit-learn==1.3.2
numpy==1.24.3
POT==0.9.5
Python==3.8.20